requirements.txt
streamlit==1.35.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
requests==2.31.0
python-whois==0.9.4
tldextract==5.1.2
fpdf2==2.7.9
Pillow==10.3.0
joblib==1.4.2
bcrypt==4.1.3
streamlit-option-menu==0.3.13
